# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ServRedi FSM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /customers/{id}/jobs:
    get:
      operationId: getCustomerJobs
      tags: [customers]
      summary: Get all jobs for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"

  # ── Jobs ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        name:
          type: string
        company:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zip:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        totalJobs:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - organizationId
        - name
        - totalJobs
        - createdAt
        - updatedAt
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        company:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        notes:
          type: string
      required:
        - name
    UpdateCustomerBody:
      type: object
      properties:
        name:
          type: string
        company:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        notes:
          type: string
    Job:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        customerId:
          type: integer
        customerName:
          type: string
        technicianId:
          type: ["integer", "null"]
        technicianName:
          type: ["string", "null"]
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled, on_hold]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        scheduledDate:
          type: ["string", "null"]
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        estimatedHours:
          type: ["number", "null"]
        actualHours:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - organizationId
        - title
        - customerId
        - customerName
        - status
        - priority
        - createdAt
        - updatedAt
...